Remove databaseStream
All the functionality has been extracted from this module so it can now be
removed.

We replicate the sequential processing of the previous stream-based
implmenetation by adding the `asyncMapInSequence` helper.

This commit does not reimplement the `bulk` feature of the database client
previously used by the `_writev` mentod in the writable stream subclass.