Add start and end datetime attributes to post-ingest SNS messages
SNS message attributes do not support a native timestamp type. SNS filter
policies do not support filtering string attributes within a range. This means
that we will need to encode the datetime attribute as a number.

To allow SNS messages to be filterable by start and end date this commit adds 2
attributes to the published messages

- `startUnixEpochMsOffset`
- `endUnixEpochMsOffset`

The values of these attributes will be the number of seconds since (or before)
the UNIX epoch

Since the STAC spec allows for collections to have null values for the temporal
extent and SNS numeric message attributes cannot have an empty value we exclude
the attributes if we do not have non-null values for them.